refactor: define payment types
